Hi genebuilders (mainly),
I'm making a small change to the way the master_% tables on the
production database are used.

Currently, each master_% table (e.g., master_external_db) is a verbatim
copy that should be exactly what goes into each Core-like database
(indexes and all).  This creates a problem in that we can't remove
entries very easily.  Removing an entry from the master_% table would
not remove it from any of the other Core-like databases when the syncing
script is run.  The script would just assume that the entry is new
in the Core database and suggest SQL for re-inserting it into the
production database.

So, I will add a boolean field, is_current, to each master_% table, and
then create a view for each master table with the name of the Core table
without the extra column.  So, for master_external_db, there will be a
view in the production database called external_db.  This view holds
exactly the data that should go into the Core-like database.

Talking to Amonida, it seems the genebuilders have recently decided to
get the data from these tables using "SELECT *" and then load it with
"LOAD DATA FROM INFILE", which is good.  This is would exactly be what
would be needed to be done to get the data from the view (as a regular
mysqldump doesn't dump data from views).

Just FYI.


Andreas