Add source to XML and modify the source based on the annotation
By default the source is havana. Because we are working on NoMerge,
the source in the Loutre database can be ensembl_havana, havana or
ensembl. When an annotator modify an ensembl transcript, it becomes
an ensembl_havana transcript and the gene also has the ensembl_havana
source.

There is some hack for the client so it send a message to Zmap so the
transcript is redrawn is the correct track